fixed weird float/double numbers issue + scaling

+++ Assets/Scripts/EnemyBehaviour/Enemy.cs
@@ -24,7 +24,7 @@ namespace EnemyBehaviour
 
         public float EnergyDrop
         {
-            get => energyDrop;
+            get => (int)energyDrop;
             set => energyDrop = value;
         }
